Constrain reduced to unsigned integers
